backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
pydantic>=2.6.1
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
python-multipart>=0.0.9
pdfminer.six>=20221105
httpx>=0.27.0
python-dotenv>=1.0.0
alembic>=1.13.1
structlog>=24.1.0
slowapi>=0.1.8
starlette>=0.35.0
psycopg[binary]>=3.1.18
psycopg2-binary>=2.9.9
sentry-sdk[fastapi]>=1.40.0
